# server/doc/api.yaml
openapi: 3.0.0
info:
  title: Notes Service
  description: Notes Service
  version: 1.0.0
servers:
  - url: /
components:
  schemas:
...
    Note:
      required:
        - title
      properties:
        id:
          type: string
        title:
          type: string
        order:
          type: number
        description:
          type: string
        isUrgent:
          type: boolean
        userId:
          type: string
        categoryId:
          type: string
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /categories/{categoryId}/notes:
    parameters:
      - name: categoryId
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Notes
      summary: Get all notes by categoryId
      description: Gets notes by the Category ID
        (e.g. “/category/1/notes)
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
        401:
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Notes
      summary: Create new note
      description: Creates a new note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                title:
                  type: string
                  description: The note's title
                order:
                  type: integer
                  description: The note's order
                description:
                  type: string
                isUrgent:
                  type: boolean
                  description: Is note urgent
                userId:
                  type: string
                  description: The note's Assignee
                categoryId:
                  type: string
                  description: The note's category
              required:
                - title
                - order
                - isUrgent
      responses:
        201:
          description: The note has been created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        400:
          description: Bad request
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /categories/{categoryId}/notes/{noteId}:
    parameters:
      - name: categoryId
        in: path
        required: true
        schema:
          type: string
      - name: noteId
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Notes
      summary: Get Note by categoryId and noteId
      description: Gets the Note by the Category's and note ID
        (e.g. “/category/1/notes/123”)
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        401:
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags:
        - Notes
      summary: Update Note
      description: Updates a Note by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                title:
                  type: string
                  description: The note's title
                order:
                  type: integer
                  description: The note's order
                description:
                  type: string
                isUrgent:
                  type: boolean
                  description: Is note urgent
                userId:
                  type: string
                  description: The note's Assignee
                categoryId:
                  type: string
                  description: The note's Category
              required:
                - order
                - userId
                - categoryId
      responses:
        200:
          description: The note has been updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        400:
          description: Bad request
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: Note not found
    delete:
      tags:
        - Notes
      summary: Delete Note
      description: Deletes a Note by ID.
      responses:
        204:
          description: The note has been deleted
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: Note not found
...